+++ main.js
@@ -1,6 +1,7 @@
 import GameOver from "./public/assets/scenes/GameOver.js";
 import Juego from "./public/assets/scenes/Juego.js";
 import Juego2 from "./public/assets/scenes/Juego2.js";
+import Juego3 from "./public/assets/scenes/Juego3.js";
 
 
 // Create a new Phaser config object
@@ -30,7 +31,7 @@ const config = {
   // List of scenes to load
   // Only the first scene will be shown
   // Remember to import the scene before adding it to the list
-  scene: [Juego, Juego2, GameOver],
+  scene: [Juego, Juego2, Juego3,GameOver],
 };
 
 // Create a new Phaser game instance
