+++ public/tilemaps/nivel3.json
@@ -54,6 +54,12 @@
          "id":2,
          "name":"plataformas",
          "opacity":1,
+         "properties":[
+                {
+                 "name":"colision",
+                 "type":"bool",
+                 "value":true
+                }],
          "type":"tilelayer",
          "visible":true,
          "width":48,
